+++ src/main/java/flab/delideli/controller/UpdateMemberController.java
@@ -8,11 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpSession;
 
@@ -26,14 +22,14 @@ public class UpdateMemberController {
     private static final ResponseEntity OK_RESPONSE_ENTITY = new ResponseEntity(HttpStatus.OK);
     private static final String USER_ID = "USER_ID";
 
-    @RequestMapping(value = "/{userId}", method = RequestMethod.PATCH)
+    @PatchMapping(value = "/{userId}")
     public ResponseEntity updateUser(HttpSession session, @RequestBody UpdateDTO updateDTO) {
         String currentUserId = (String) session.getAttribute(USER_ID);
         memberService.updateUserInfo(currentUserId ,updateDTO);
         return OK_RESPONSE_ENTITY;
     }
 
-    @RequestMapping(value="/{userId}", method = RequestMethod.DELETE)
+    @DeleteMapping(value="/{userId}")
     public ResponseEntity deleteUser(HttpSession session) {
         String currentUserId = (String) session.getAttribute(USER_ID);
         memberService.deleteUserInfo(currentUserId);
