+++ src/main/java/flab/delideli/interceptor/LoginInterceptor.java
@@ -1,34 +1,15 @@
 package flab.delideli.interceptor;
 
-import flab.delideli.exception.UnauthorizedException;
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 
-@Component
-public class LoginInterceptor implements HandlerInterceptor {
+public interface LoginInterceptor {
 
-    private static final String USER_ID = "USER_ID";
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception;
 
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        HttpSession session = request.getSession();
-        String currentUserId = (String) session.getAttribute("USER_ID");
+    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception;
 
-        if (currentUserId != null)
-            return true;
-        throw new UnauthorizedException();
-    }
-
-    @Override
-    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
-    }
-
-    @Override
-    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
-    }
+    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception;
 }
